Version 0.06
New command signature
     Mkcolorscheme[!] [scheme-name]

If invoked with a ! then the generated colorscheme file is automatically written to ~/.vim/colors/<scheme-name>.vim if possible.  Only works with has("unix") for now, feel free to fix this.

The included :command is no longer commented out and properly invokes the Mkcolorscheme() function with the <bang> syntax.